flask_app/requirements.txt
catboost==1.2.7
Flask==3.0.0
flask-babel==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-Caching==2.3.0
Flask-WTF==1.2.1
Jinja2==3.1.3
jmespath==1.0.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
matplotlib==3.9.2
matplotlib-inline==0.1.6
mistune==3.0.2
modules==1.0.0
numpy==1.26.4
packaging==23.2
pandas==2.2.3
parsedatetime==2.6
plotly==5.24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.25
w3lib==2.1.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wsproto==1.2.0
WTForms==3.1.2
